Moved all ctor logic to the pipeline provider
