+++ pkg/logs/client/destination.go
@@ -1,6 +1,6 @@
 package client
 
-// Destination TODO
+// Destination sends a payload to a specific endpoint over a given network protocol.
 type Destination interface {
 	Send(payload []byte) error
 	SendAsync(payload []byte)
