+++ pkg/logs/client/destinations.go
@@ -5,43 +5,14 @@
 
 package client
 
-import (
-	"github.com/DataDog/datadog-agent/pkg/logs/client/http"
-	"github.com/DataDog/datadog-agent/pkg/logs/client/tcp"
-	"github.com/DataDog/datadog-agent/pkg/logs/config"
-)
-
 // Destinations holds the main destination and additional ones to send logs to.
 type Destinations struct {
 	Main        Destination
 	Additionals []Destination
 }
 
-// NewDestinations takes endpoints configuration and returns
-func NewDestinations(endpoints *config.Endpoints, destinationsContext *tcp.DestinationsContext) *Destinations {
-	destinations := &Destinations{}
-	var additionals []Destination
-
-	if endpoints.UseHTTP {
-		destinations.Main = http.NewDestination(endpoints.Main)
-		for _, endpoint := range endpoints.Additionals {
-			additionals = append(additionals, http.NewDestination(endpoint))
-		}
-		destinations.Additionals = additionals
-	} else {
-		destinations.Main = tcp.NewDestination(endpoints.Main, endpoints.UseProto, destinationsContext)
-		for _, endpoint := range endpoints.Additionals {
-			additionals = append(additionals, tcp.NewDestination(endpoint, endpoints.UseProto, destinationsContext))
-		}
-		destinations.Additionals = additionals
-	}
-
-	return destinations
-}
-
-// NewDestinationsOld returns a new destinations composite.
-// TODO(achntrl): To remove
-func NewDestinationsOld(main Destination, additionals []Destination) *Destinations {
+// NewDestinations returns a new destinations composite.
+func NewDestinations(main Destination, additionals []Destination) *Destinations {
 	return &Destinations{
 		Main:        main,
 		Additionals: additionals,
