+++ pkg/logs/client/destinations_context_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-package tcp
+package client
 
 import (
 	"github.com/stretchr/testify/assert"
