+++ pkg/logs/sender/sender.go
@@ -5,210 +5,8 @@
 
 package sender
 
-import (
-	"context"
-	"time"
-
-	"github.com/DataDog/datadog-agent/pkg/logs/client"
-	"github.com/DataDog/datadog-agent/pkg/logs/client/tcp"
-	"github.com/DataDog/datadog-agent/pkg/logs/message"
-	"github.com/DataDog/datadog-agent/pkg/logs/metrics"
-)
-
-// Sender sends logs to different destination and reports to auditor
+// Sender sends logs to different destinations.
 type Sender interface {
 	Start()
 	Stop()
-}
-
-// SingleSender sends single log to different destinations.
-type SingleSender struct {
-	inputChan    chan *message.Message
-	outputChan   chan *message.Message
-	destinations *client.Destinations
-	done         chan struct{}
-}
-
-// NewSingleSender returns an new SingleSender.
-func NewSingleSender(inputChan, outputChan chan *message.Message, destinations *client.Destinations) *SingleSender {
-	return &SingleSender{
-		inputChan:    inputChan,
-		outputChan:   outputChan,
-		destinations: destinations,
-		done:         make(chan struct{}),
-	}
-}
-
-// Start starts the SingleSender
-func (s *SingleSender) Start() {
-	go s.run()
-}
-
-// Stop stops the SingleSender,
-// this call blocks until inputChan is flushed
-func (s *SingleSender) Stop() {
-	close(s.inputChan)
-	<-s.done
-}
-
-// run lets the SingleSender send messages.
-func (s *SingleSender) run() {
-	defer func() {
-		s.done <- struct{}{}
-	}()
-
-	for payload := range s.inputChan {
-		s.send(payload)
-	}
-}
-
-// send keeps trying to send the message to the main destination until it succeeds
-// and try to send the message to the additional destinations only once.
-func (s *SingleSender) send(payload *message.Message) {
-	for {
-		// this call is blocking until payload is sent (or the connection destination context cancelled)
-		err := s.destinations.Main.Send(payload.Content)
-		if err != nil {
-			if err == context.Canceled {
-				metrics.DestinationErrors.Add(1)
-				// the context was cancelled, agent is stopping non-gracefully.
-				// drop the message
-				break
-			}
-			switch err.(type) {
-			case *tcp.FramingError:
-				metrics.DestinationErrors.Add(1)
-				// the message can not be framed properly,
-				// drop the message
-				break
-			default:
-				metrics.DestinationErrors.Add(1)
-				// retry as the error can be related to network issues
-				continue
-			}
-		}
-		for _, destination := range s.destinations.Additionals {
-			// send to a queue then send asynchronously for additional endpoints,
-			// it will drop messages if the queue is full
-			destination.SendAsync(payload.Content)
-		}
-
-		metrics.LogsSent.Add(1)
-		break
-	}
-	s.outputChan <- payload
-}
-
-// BatchSender is responsible for sending logs to different destinations in batches
-type BatchSender struct {
-	inputChan     chan *message.Message
-	outputChan    chan *message.Message
-	destinations  *client.Destinations
-	done          chan struct{}
-	batchTimeout  time.Duration
-	messageBuffer *MessageBuffer
-}
-
-// NewBatchSender returns an new BatchSender.
-func NewBatchSender(inputChan, outputChan chan *message.Message, destinations *client.Destinations) *BatchSender {
-	return &BatchSender{
-		inputChan:     inputChan,
-		outputChan:    outputChan,
-		destinations:  destinations,
-		done:          make(chan struct{}),
-		batchTimeout:  5 * time.Second,               // TODO Change this to parameter
-		messageBuffer: NewMessageBuffer(20, 2000000), // TODO change this to parameter
-	}
-}
-
-// Start starts the BatchSender
-func (b *BatchSender) Start() {
-	go b.run()
-}
-
-// Stop stops the BatchSender,
-// this call blocks until inputChan is flushed
-func (b *BatchSender) Stop() {
-	close(b.inputChan)
-	<-b.done
-}
-
-// run lets the BatchSender send messages.
-func (b *BatchSender) run() {
-	flushTimer := time.NewTimer(b.batchTimeout)
-	defer func() {
-		flushTimer.Stop()
-		b.done <- struct{}{}
-	}()
-
-	for {
-		select {
-		case payload, isOpen := <-b.inputChan:
-			if !isOpen {
-				// inputChan has been closed, no more payload are expected
-				// TODO remember to flush your buffer!!
-				return
-			}
-			sendNow, appendAgain := b.messageBuffer.tryAppendToBuffer(payload)
-			if sendNow {
-				// message buffer is full, either reaching maxBatchCount of maxRequestSize
-				// send request now. reset the timer
-				if !flushTimer.Stop() {
-					<-flushTimer.C
-				}
-				b.sendBuffer()
-				flushTimer.Reset(b.batchTimeout)
-			}
-			if appendAgain {
-				// it's possible we didn't append last try because maxRequestSize is reached
-				// append it again after the sendbuffer is flushed
-				b.messageBuffer.tryAppendToBuffer(payload)
-			}
-		case <-flushTimer.C:
-			// the timout expired, the content is ready to be sent
-			b.sendBuffer()
-			flushTimer.Reset(b.batchTimeout)
-		}
-	}
-}
-
-// send keeps trying to send the message to the main destination until it succeeds
-// and try to send the message to the additional destinations only once.
-func (b *BatchSender) sendBuffer() {
-	if b.messageBuffer.isEmpty() {
-		return
-	}
-
-	batchedContent := b.messageBuffer.getByteBuffer()
-	//log.Info("MAOMAO " + string(batchedContent[12:20]))
-
-	for {
-		// this call is blocking until payload is sent (or the connection destination context cancelled)
-		err := b.destinations.Main.Send(batchedContent)
-		if err != nil {
-			if err == context.Canceled {
-				metrics.DestinationErrors.Add(1)
-				// the context was cancelled, agent is stopping non-gracefully.
-				// drop the message
-				break
-			}
-			switch err.(type) { //TODO think that's the possible error
-			default:
-				metrics.DestinationErrors.Add(1)
-				// retry as the error can be related to network issues
-				continue
-			}
-		}
-		for _, destination := range b.destinations.Additionals {
-			// TODO this does nothing right now
-			destination.SendAsync(batchedContent)
-		}
-
-		metrics.LogsSent.Add(1)
-		break
-	}
-	for _, m := range b.messageBuffer.messageBuffer {
-		b.outputChan <- m
-	}
-	b.messageBuffer.clear()
 }
